Refactored scripts to individual files for readability.
- faster script builds.
